Unify buildRoll's skipUI true/false paths
The skipUI flag previously marked a fork in the codepath. Both paths
were very similar. This commit unifies those paths, so that the skipUI
flag _only_ skips displaying the user the roll UI. It instead provides
default values and carries on down the same code path.

The major difference was that the skipUI=true + partless path did not
attempt to replace the `<damageSection>` magic string, but I'm pretty
sure that was an oversight anyway.