Pass to RollTree.populate() by parameter instead of by field
A small change to make data-flow more explicit. The intended API now is
that `RollTree.populate()` is _entirely_ responsible for setting the
object's fields, and none (but those set in the constructor) should be
used prior to calling `populate()`.

This change is made more reasonable by the fact that `this.formula` and
`this.contexts` were never used after calling `populate()`.