Stop pretending anyone cares about the results of `RollTree.buildRoll`
I found exactly one segment of code checking the data passed back up
from `buildRoll`, and that was `ItemSFRPG.rollConsumable` wanting to
know if it should consume the item or not. It did this by checking if
the returned data had a truthy `callbackResult` field. Considering that
this part of the code had no knowledge of the callback, it was better to
just pass a success/fail boolean back up instead of all that data.

The large footprint of this change is a consequence of me ensuring that
there are explicit breaks of that data dependency up the call-stack from
`buildRoll()`. I got a bit aggressive making sure the upstream functions
returned a `Promise<void>` wherever possible.