Remove `<damageSection>` magic string.
Previous implementation used `<damageSection>` as a placeholder that
laters parts of the roll code would substitute for the part formula.
This commit changes the semantics such that whenever `RollTree` has
`this.options.parts`, the formula passed to `buildRoll()` instead
represents the bonuses that will be applied to any part that is a
primary section.

A considered alternative was to replace `<damageSection>` with a proper
attribute, such as `@additional.sections.primary`, but this was
considered to be a substantial change and effort for marginal value.