Convert `RollTree.prototype.buildRoll` to static `RollTree.buildRoll`
Instances of `RollTree` are only used by the roll dialog, and their
usefulness is largely over after `buildRoll` is called. In theory they
could have been reused, but in practice they are so stateful that new
`RollTree` objects were created for each invocation of `buildRoll`.
Converting `buildRoll` to a static method that creates a local
`RollTree` object captures how the class was being used in practice.

Along side this change `RollTree.prototype.populate` was changed to take
the context as a parameter, and the context is no longer stored on the
tree. The intent of this was to make the relationship between `populate`
and the `RollContext` more explicit, instead of by implication, when
reading `RollDialog.prototype._onSelectorChanged`, which changes the
context then re-populates the stored `RollTree` object.